requirements.txt
asgiref==3.4.1
async-timeout==4.0.2
baidu-aip==4.16.3
certifi==2022.5.18.1
chardet==4.0.0
charset-normalizer==2.0.12
Deprecated==1.2.13
Django==3.2.13
djangorestframework==3.13.1
idna==3.3
importlib-metadata==4.8.3
packaging==21.3
PyJWT==2.4.0
PyMySQL==1.0.2
pyparsing==3.0.7
pytz==2022.1
redis==4.3.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.1.1
urllib3==1.26.9
wrapt==1.14.1
zipp==3.6.0
